requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
greenlet==2.0.1
jdatetime==4.1.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.24.0
pandas==1.5.2
protobuf==3.20.1
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==2.0.0
sqlparse==0.4.3
typing_extensions==4.4.0
